Moves all of the layer related functionality to a new storylayers app.
