+++ mapstory/import_handlers.py
@@ -3,7 +3,7 @@ from django import db
 from osgeo_importer.handlers import ImportHandlerMixin, ensure_can_run
 from osgeo_importer.inspectors import OGRTruncatedConverter
 
-from .views import layer_append_minimal
+from storylayers.views import layer_append_minimal
 
 
 class LayerAppendHandler(ImportHandlerMixin):
